DROID-2079 move date parser
